neuralsentinel_app/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
pandas==3.0.0
tensorflow==2.18.0
torch>=2.2.0
Pillow>=10.0.0
matplotlib==3.7.5        # For plotting
scikit-learn==1.8.0      # For NeuralStrength
seaborn==0.13.2          # For NeuralStrength
tqdm==4.66.4             # For NeuralStrength
umap-learn==0.5.11       # For NeuralStrength
scipy==1.10.1            # For NeuralStrength
scikit-image==0.21.0     # For NeuralStrength
foolbox==3.3.2
